### 2020-06-07 #2294 @whedon
[PRE REVIEW]: tidyfst: Tidy Verbs for Fast Data Manipulation
**Submitting author:** @hope-data-science (<a href="http://orcid.org/0000-0002-4151-3764">Tian-Yuan Huang</a>)
**Repository:** <a href="https://github.com/hope-data-science/tidyfst" target ="_blank">https://github.com/hope-data-science/tidyfst</a>
**Version:** v0.9.7
