### 2020-06-07 #2294 @arfon
> @hope-data-science it looks like the URL for the software repo does not load. Can you provide the correct URL?

I updated this.
