### 2020-06-07 #2294 @arfon
@hope-data-science - looks like there's some kind of encoding issue with your files here. Could you make sure you can get your paper to compile with this service while we're waiting to assign an editor: http://whedon.theoj.org
