### 2020-06-08 #2294 @hope-data-science
@arfon Thank you for the prompt feedback. I've try to compile this link (<https://github.com/hope-data-science/tidyfst/tree/master/docs/JOSS>) and successed, but when I click and try to download the preview, it failed. Am I supposed to change anything in the submission system or make adjustment on the files in the Github repo?

A Word version is also available here: <https://github.com/hope-data-science/tidyfst/blob/master/docs/tidyfst_JOSS.docx>. If there's any more questions, let me know.

Thanks.
