### 2020-06-11 #2294 @hope-data-science
Thank you for the feedback. I have no idea how to advance the process at the current state.
Best.

BTW: A copy of the manuscript in MS WORD could be found at <https://github.com/hope-data-science/tidyfst/blob/master/docs/tidyfst_JOSS.docx>. I tried to compile files at <https://github.com/hope-data-science/tidyfst/tree/master/docs/JOSS> but failed to the PDF. Any hints?
