### 2020-06-11 #2294 @dvanic
@hope-data-science You have an encoding issue. One possibility is that the "ç" character you have in your markdown is causing pandoc to flip - perhaps replace it with a c for now to test?

If not that one, I'd look at other things like quotes, symbols (@#brackets etc) that could be breaking pandoc.
