### 2020-06-11 #2294 @hope-data-science
@dvanic Thanks, I've made this change in the manuscript. Hope it could help later.
