### 2020-06-14 #2294 @labarba
@hope-data-science — you can't have spaces in labels of bibles entries; please fix:
https://github.com/hope-data-science/tidyfst/blob/dc57addcb3629a57c33e874544809acdc869089d/docs/JOSS/paper.bib#L64
