+++ docs/JOSS/paper.bib
@@ -61,7 +61,7 @@ RinkerKurkiewicz-485,
 
 
 @misc{
-R Core Team-479,
+R-Core-Team-479,
    Author = {R Core Team},
    Title = {R: A Language and Environment for Statistical Computing},
    Year = {2020} }
