+++ docs/JOSS/paper.md
@@ -30,7 +30,7 @@ bibliography: paper.bib
 
 # Summary
 
-The tidyfst package[@Huang-488] is an R package[@R Core Team-479] for fast data manipulation in tidy syntax. The top-level design is inherited from tidy data structure proposed by Hadley Wickham[@Wickham-458], which are: (1) each variable is a column; (2) each observation is a row; (3) each type of observational unit is a table. Moreover, the function names as well as parameter settings are very much borrowed from dplyr[@WickhamFrancois-481] and tidyr[@WickhamHenry-490], reducing the learning cost for tidyverse[@WickhamAverick-493] users. At the bottom, tidyfst is backed by the high performance package data.table[@DowleSrinivasan-480], which is speedy, stable (with little dependency), memory efficient and feature rich.
+The tidyfst package[@Huang-488] is an R package[@R-Core-Team-479] for fast data manipulation in tidy syntax. The top-level design is inherited from tidy data structure proposed by Hadley Wickham[@Wickham-458], which are: (1) each variable is a column; (2) each observation is a row; (3) each type of observational unit is a table. Moreover, the function names as well as parameter settings are very much borrowed from dplyr[@WickhamFrancois-481] and tidyr[@WickhamHenry-490], reducing the learning cost for tidyverse[@WickhamAverick-493] users. At the bottom, tidyfst is backed by the high performance package data.table[@DowleSrinivasan-480], which is speedy, stable (with little dependency), memory efficient and feature rich.
 
 Sharing similar goals, both data.table and dplyr have gained much popularity among R users. Their features have been compared widely in the community, with pros and cons suggested in ideas and tested in examples (e.g. <https://stackoverflow.com/questions/21435339/data-table-vs-dplyr-can-one-do-something-well-the-other-cant-or-does-poorly/27840349#27840349>). While some opinions might be objective, consensus could be made on at least two points: (1) data.table could handle data manipulation in less time than dplyr; (2) dplyr has a possibly more user-friendly syntax for learning and communication than data.table. The tidyfst package is designed to combine the merits of dplyr and data.table, so as to provide a suite of tidy verbs for fast data manipulation.
 
