+++ R/count_dt.R
@@ -5,7 +5,7 @@
 #'
 #' @param .data data.table/data.frame data.frame will be automatically converted
 #'   to data.table.
-#' @param ... variables to group by.
+#' @param ... Variables to group by, could receive what `select_dt` receives.
 #' @param sort logical. If TRUE result will be sorted in desending order by resulting variable.
 #' @param .name character. Name of resulting variable. Default uses "n".
 #'
@@ -18,15 +18,18 @@
 #' iris %>% add_count_dt(Species,.name = "N")
 #'
 #' mtcars %>% count_dt(cyl,vs)
+#' mtcars %>% count_dt("cyl|vs")
 #' mtcars %>% count_dt(cyl,vs,.name = "N",sort = FALSE)
 #' mtcars %>% add_count_dt(cyl,vs)
+#' mtcars %>% add_count_dt("cyl|vs")
 #'
 #' @rdname count
 #' @export
 
 count_dt = function(.data,...,sort = TRUE,.name = "n"){
   dt = as_dt(.data)
-  dot_string = substitute(list(...))
+  # dot_string = substitute(list(...))
+  dt[0] %>% select_dt(...) %>% names() -> dot_string
   if(sort)
     eval(parse(text =
                  str_glue("dt[,.({.name} = .N),by = dot_string][order(-{.name})]")))
@@ -37,11 +40,14 @@ count_dt = function(.data,...,sort = TRUE,.name = "n"){
 #' @rdname count
 #' @export
 add_count_dt = function(.data,...,.name = "n"){
-  #dt = as_dt(.data)
   dt = as.data.table(.data)
-  dot_string = substitute(list(...))
+  # dot_string = substitute(list(...))
+  dt[0] %>% select_dt(...) %>% names() -> dot_string
   dt[,(.name):=.N,by = dot_string][]
 }
+
+
+
 
 
 
