+++ R/slice.R
@@ -32,6 +32,8 @@
 #' a = iris
 #' slice_dt(a,1,2)
 #' slice_dt(a,2:3)
+#' slice_dt(a,141:.N)
+#' slice_dt(a,1,.N)
 #' slice_head_dt(a,5)
 #' slice_head_dt(a,0.1)
 #' slice_tail_dt(a,5)
@@ -47,8 +49,7 @@
 #' @rdname slice
 #' @export
 slice_dt = function(.data,...){
-  .data = as_dt(.data)
-  .data[c(...)]
+  as_dt(.data)[eval(substitute(c(...)))]
 }
 
 #' @rdname slice
