+++ NAMESPACE
@@ -60,6 +60,7 @@ export(pull_dt)
 export(rbindlist)
 export(relocate_dt)
 export(rename_dt)
+export(rename_with_dt)
 export(replace_dt)
 export(replace_na_dt)
 export(right_join_dt)
