### 2020-06-17 #2294 @hope-data-science
TODO list:
1. Miss the citation of the package fst, add it in next correction.
2. Cite "R Core Team" correctly.
