### 2020-06-20 #2294 @hope-data-science
@mikldk Perhaps someone from this list: <https://github.com/hope-data-science/tidyfst/stargazers>. I think they might have the first-hand experience of using this package and make comments on it to make it better.
