### 2020-06-20 #2294 @hope-data-science
Also, I recommend this folk (https://github.com/Atrebas). He has written a great post (from which I benefited from) on the related topic at <https://atrebas.github.io/post/2019-03-03-datatable-dplyr/>. I think he might be interested and give great suggestions.
