### 2020-06-26 hope-data-science/tidyfst#4 @rcannood
[JOSS review] Remarks
Hello Tian-Yuan Huang!

I am reviewing the tidyfst manuscript for your submission at JOSS (openjournals/joss-reviews#2388). Below are some remarks regarding the manuscript and software repository. Please complete the tasks below and tick off when done.

Robrecht

**Performance: If there are any performance claims of the software, have they been confirmed?**

The manuscript reads: "While some opinions might be objective, consensus could be made on at least two points: (1) data.table could handle data manipulation in less time than dplyr; (2) dplyr has a possibly more user-friendly syntax for learning and communication than data.table. The tidyfst package is designed to combine the merits of dplyr and data.table, so as to provide a suite of tidy verbs for fast data manipulation."

- [x] Please include a benchmark comparing dplyr, data.table and tidyfst using, for example, [microbenchmark](https://adv-r.hadley.nz/perf-measure.html#microbenchmarking). You can include the benchmark either in the manuscript or in a separate vignette.

**Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?**

Looking at the documentation of tidyfst on CRAN, I see a lot of functions for which the documentation simply refers to dplyr, e.g. for `tidyfst::mutate_dt()`: "Analogous function for mutate and transmute in dplyr". The documentation of this function can reference to dplyr, but should be able to stand on its own.

- [x] Please write descriptions for each of the functions such that its functionality can be understood.

**Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?**

- [x] Please write [unit tests](http://r-pkgs.had.co.nz/tests.html) for each of the functions.

**Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support**

- [x] Please include these sets of guidelines. [`usethis::use_tidy_github()`](https://cran.r-project.org/web/packages/usethis/usethis.pdf#Rfn.use.Rul.tidy.Rul.github.Rul.actions.1) sets up templates for each of these guidelines.

**References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?**

- [x] The reference "Team, R. C." is not a person, should be "R Core Team". Note that this manuscript is from 2013, not 2020. Please fix.
- [x] References are not complete: please include the journal, DOI, publication date.
- [x] Please include a space before each reference.
- [x] Several references are not papers but are probably software repositories. Please specify that it is an R package and include a URL. A good example is the tidyverse paper: [source](https://github.com/tidyverse/tidyverse/tree/master/vignettes), [pdf](https://joss.theoj.org/papers/10.21105/joss.00388).
