### 2020-06-26 #2388 @rcannood
@mikldk: I asked @hope-data-science to write unit tests and provide documentation for each of the functions in tidyfst. Since there are quite a lot of functions, I assume this will take a while.
