### 2020-06-30 c5b78ef
Delete all_at_if.R
Deprecated
