### 2020-06-30 a077598
Update count_dt.R
Update description.
