### 2020-06-30 640b9ad
Update count_dt.R
