### 2020-06-30 cea2d9c
Update distinct_dt.R
