+++ R/filter_dt.R
@@ -1,6 +1,6 @@
 
 #' @title Filter entries in data.frame
-#' @description Analogous function for \code{filter} in \pkg{dplyr}.
+#' @description choose rows where conditions are true. Analogous function for \code{filter} in \pkg{dplyr}.
 #' @param .data data.frame
 #' @param ... List of variables or name-value pairs of summary/modifications
 #'   functions.
