### 2020-07-01 484bbc7
Update count_dt.R
