+++ R/count_dt.R
@@ -2,7 +2,6 @@
 
 #' @title Count observations by group
 #' @description Count the unique values of one or more variables.
-#' @description Analogous function for \code{count} and \code{add_count} in \pkg{dplyr}.
 #' @param .data data.table/data.frame data.frame will be automatically converted
 #'   to data.table.
 #' @param ... Variables to group by, could receive what `select_dt` receives.
