+++ R/filter_dt.R
@@ -1,7 +1,6 @@
 
 #' @title Filter entries in data.frame
 #' @description Choose rows where conditions are true.
-#' Analogous function for \code{filter} in \pkg{dplyr}.
 #' @param .data data.frame
 #' @param ... List of variables or name-value pairs of summary/modifications
 #'   functions.
