### 2020-07-01 f7fe6b8
Update fst.R
