+++ R/group_by.R
@@ -1,5 +1,6 @@
 
 #' @title Group by variable(s) and implement operations
+#' @description Carry out data manipulation within specified groups.
 #' @description Using \code{setkey} and \code{setkeyv} in \pkg{data.table}
 #' to carry out \code{group_by}-like functionalities in \pkg{dplyr}. This is
 #' not only convenient but also efficient in computation.
