### 2020-07-02 dc0d3d7
Delete join_dt.R
