+++ R/join_dt.R
@@ -1,271 +1,0 @@
-#' @title Join table by common keys
-#' @description Join operations.
-#' @name join
-#' @param x data.frame
-#' @param y data.frame
-#' @param by
-#'   For functions with suffix \code{_dt}.
-#'    A character vector of variables to join by. If NULL, the default,
-#'   *_join() will do a natural join, using all variables with common names
-#'   across the two tables. A message lists the variables so that you can check
-#'   they're right (to suppress the message, simply explicitly list the
-#'   variables that you want to join). To join by different variables on x and y
-#'   use a named vector. For example, by = c("a" = "b") will match x.a to y.b.
-#' @param by For functions without suffix \code{_dt},
-#'   this parameter will pass to the
-#'   \code{on} paramter to the data.table. Details could be found at
-#'   \code{\link[data.table]{data.table}}.Examples included:
-#'   1.\code{.by = c("a","b")} (this is a must for \code{set_full_join});
-#'   2.\code{.by = c(x1="y1", x2="y2")};
-#'   3.\code{.by = c("x1==y1", "x2==y2")};
-#'   4.\code{.by = c("a", V2="b")};
-#'   5.\code{.by = .(a, b)};
-#'   6.\code{.by = c("x>=a", "y<=b")} or \code{.by = .(x>=a, y<=b)}.
-#' @param suffix If there are non-joined duplicate variables in x and y, these
-#'   suffixes will be added to the output to disambiguate them. Should be a
-#'   character vector of length 2.
-#' @description For functions without suffix \code{_dt}, they use
-#' \code{X[Y]} syntax to join tables, and pass the "by" parameter to
-#' "on" in data.table. They have different features and syntax. They
-#' starts with first letter of left/right/full/inner/semi/anti.
-
-#' @return data.table
-#' @seealso \code{\link[dplyr]{left_join}}
-#' @examples
-#'
-#' workers = fread("
-#'     name company
-#'     Nick Acme
-#'     John Ajax
-#'     Daniela Ajax
-#' ")
-#'
-#' positions = fread("
-#'     name position
-#'     John designer
-#'     Daniela engineer
-#'     Cathie manager
-#' ")
-#'
-#' workers %>% inner_join_dt(positions)
-#' workers %>% left_join_dt(positions)
-#' workers %>% right_join_dt(positions)
-#' workers %>% full_join_dt(positions)
-#'
-#' # filtering joins
-#' workers %>% anti_join_dt(positions)
-#' workers %>% semi_join_dt(positions)
-#'
-#' # To suppress the message, supply 'by' argument
-#' workers %>% left_join_dt(positions, by = "name")
-#'
-#' # Use a named 'by' if the join variables have different names
-#' positions2 = setNames(positions, c("worker", "position")) # rename first column in 'positions'
-#' workers %>% inner_join_dt(positions2, by = c("name" = "worker"))
-#'
-#' # the syntax of non-"_dt" suffix joins is a bit different
[…]
