### 2020-07-02 7432a5c
Update mutate_when.R
