### 2020-07-02 6b104b2
Update relocate_dt.R
