### 2020-07-08 002a371
Update sample.R
