### 2020-07-08 b821c76
Update sample.R
