### 2020-07-08 86d8e43
Update separate.R
