+++ R/set.R
@@ -1,163 +1,0 @@
-
-#' @title Fast operations of data.table by reference (II)
-#' @description Combination of \code{set*} functions provided by \pkg{data.table}.
-#'  This is memeroy efficient because no copy is made at all.
-#' @param .data A data.table
-#' @param ...
-#'  For \code{set_in_dt}: Receive '[i,j,by]' in data.table syntax.
-#'   See \code{\link[tidyfst]{in_dt}}.
-#'  For \code{set_mutate}:
-#'   List of variables or name-value pairs of summary/modifications functions.
-#'   See \code{\link[tidyfst]{mutate_dt}}.
-#'  For \code{set_arrange}:
-#'   Arrange by what group? Minus means descending order.
-#'   See \code{\link[tidyfst]{arrange_dt}}.
-#'  For \code{set_rename}:
-#'   List of variables or name-value pairs of summary/modifications functions.
-#'   See \code{\link[tidyfst]{rename_dt}}.
-#'  For \code{set_relocate}: Columns to move.
-#'   See \code{\link[tidyfst]{relocate_dt}}
-#'  For \code{set_add_count}: Variables to group by.
-#'   See \code{\link[tidyfst]{add_count_dt}}
-#'  For \code{set_replace_na} and \code{set_fill_na}:
-#'   Colunms to be replaced or filled. If not specified, use all columns.
-#'   See \code{\link[tidyfst]{fill_na_dt}}
-#' @param by 	Mutate by which group(s)?
-#' @param cols For \code{set_arrange}:A character vector of column names of x by which to order.
-#' If present, override \code{...}. Defaults to \code{NULL}.
-#' @param order For \code{set_arrange}:An integer vector with only possible
-#' values of 1 and -1, corresponding to ascending and descending order.
-#' Defaults to 1.
-#' @param how For \code{set_relocate}:The mode of movement, including "first","last","after","before". Default uses "first".
-#' @param where For \code{set_relocate}:Destination of columns selected by .... Applicable for "after" and "before" mode.
-#' @param .name For \code{set_add_count}:
-#' Character. Name of resulting variable. Default uses "n".
-#' @param to What value should NA replace by?
-#' @param direction Direction in which to fill missing values. Currently either "down" (the default) or "up".
-#' @details These are a set of functions for modification
-#' on data.table by reference. They follow the same syntax of similar
-#' \pkg{tidyfst} functions. They do not return the result and considered
-#' to be memory efficient.
-#' @return  The input is modified by reference, and returned (invisibly)
-#' so it can be used in compound statements.
-#' @seealso \code{\link[tidyfst]{set_dt}}
-#' @examples
-#'
-#' library(pryr)
-#' rm(list = ls())
-#'
-#' nr_of_rows <- 1e5
-#'
-#' df <- data.table(
-#'   Logical = sample(c(TRUE, FALSE, NA),
-#'                    prob = c(0.85, 0.1, 0.05),
-#'                    nr_of_rows, replace = TRUE)
-#' )
-#'
-#' mem_change(mutate_dt(df,one = 1) -> res)
-#'
-#' # makes no copy, update on "df" directly
-#' mem_change(set_mutate(df,one = 1))
-#'
-#' all.equal(res,df)
-
-#' @rdname set
-#' @export
-set_in_dt = function(.data,...){
-  .data[...]
-}
-
-#' @rdname set
-#' @export
-set_mutate = function(.data,...,by = NULL){
-  .data = setDT(.data)
-  eval(substitute(.data[,`:=`(...),by]))
-}
-
[…]
