### 2020-07-08 ac1cdd0
Delete set_dt.R
move to tidyft already
