### 2020-07-08 dd1d2d4
Delete show_tibble.R
