### 2020-07-08 2514369
Delete slice_dt.R
