+++ R/summarise_dt.R
@@ -1,7 +1,8 @@
 
 #' @title Summarise columns to single values
 #' @name summarise_dt
-#' @description Analogous function for \code{summarise}  in \pkg{dplyr}.
+#' @description Summarise group of values into one value for each group. If there is only one group, then only one value would be returned.
+#'  The summarise function should always return a single value.
 #' @param .data data.frame
 #' @param ... List of variables or name-value pairs of summary/modifications
 #'   functions for \code{summarise_dt}.Additional parameters to be passed to
