### 2020-07-08 bf0dcf4
Update unite.R
