+++ R/slice.R
@@ -12,7 +12,7 @@
 #' @param .data A data.table
 #' @param ... Provide either positive values to keep, or negative values to drop.
 #'   The values provided must be either all positive or all negative.
-#'
+#' @param by Slice by which group(s)?
 #' @param n When larger than or equal to 1, the number of rows.
 #'   When between 0 and 1, the proportion of rows to select.
 #'
@@ -45,64 +45,205 @@
 #' slice_sample_dt(a,10)
 #' slice_sample_dt(a,0.1)
 #'
+#'
+#' # use by to slice by group
+#'
+#' ## following codes get the same results
+#' slice_dt(a,1:3,by = "Species")
+#' slice_dt(a,1:3,by = Species)
+#' slice_dt(a,1:3,by = .(Species))
+#'
+#' slice_head_dt(a,2,by = Species)
+#' slice_tail_dt(a,2,by = Species)
+#'
+#' slice_max_dt(a,Sepal.Length,3,by = Species)
+#' slice_max_dt(a,Sepal.Length,3,by = Species,with_ties = FALSE)
+#' slice_min_dt(a,Sepal.Length,3,by = Species)
+#' slice_min_dt(a,Sepal.Length,3,by = Species,with_ties = FALSE)
+#'
+#' slice_sample_dt(a,.1,by = Species)
+#' slice_sample_dt(a,3,by = Species)
+#' slice_sample_dt(a,51,replace = TRUE,by = Species)
+#'
 
 #' @rdname slice
 #' @export
-slice_dt = function(.data,...){
-  as_dt(.data)[eval(substitute(c(...)))]
+slice_dt = function(.data,...,by = NULL){
+  dt = as_dt(.data)
+  substitute(by) %>% deparse() -> by_char
+  if(by_char == "NULL") dt[eval(substitute(c(...)))]
+  else eval(substitute(
+    dt[,.SD[c(...)],by]
+  ))
 }
 
+
+# slice_dt = function(.data,...){
+#   as_dt(.data)[eval(substitute(c(...)))]
+# }
+
 #' @rdname slice
 #' @export
-slice_head_dt = function(.data,n){
+slice_head_dt = function(.data,n,by = NULL){
   .data = as_dt(.data)
-  if(abs(n) >= 1L) head(.data,n)
-  else head(.data,nrow(.data)*n)
+  substitute(by) %>% deparse() -> by_char
+  if(by_char == "NULL"){
+    if(abs(n) >= 1L) head(.data,n)
+    else head(.data,nrow(.data)*n)
+  }else{
+    if(abs(n) >= 1L)
+      eval(substitute(
+        .data[,head(.SD,n),by]
+      ))
+    else
+      eval(substitute(
+        .data[,head(.SD,nrow(.SD)*n),by]
+      ))
+  }
 }
 
+
+# slice_head_dt = function(.data,n){
+#   .data = as_dt(.data)
+#   if(abs(n) >= 1L) head(.data,n)
+#   else head(.data,nrow(.data)*n)
+# }
+
 #' @rdname slice
 #' @export
-slice_tail_dt = function(.data,n){
+slice_tail_dt = function(.data,n,by = NULL){
   .data = as_dt(.data)
-  if(abs(n) >= 1L) tail(.data,n)
-  else tail(.data,nrow(.data)*n)
+  substitute(by) %>% deparse() -> by_char
+  if(by_char == "NULL"){
+    if(abs(n) >= 1L) tail(.data,n)
+    else tail(.data,nrow(.data)*n)
+  }else{
+    if(abs(n) >= 1L)
+      eval(substitute(
+        .data[,tail(.SD,n),by]
+      ))
+    else
[…]
