+++ NAMESPACE
@@ -8,6 +8,7 @@ export(arrange_dt)
 export(as.data.table)
 export(as_dt)
 export(as_fst)
+export(between)
 export(chop_dt)
 export(col_rn)
 export(complete_dt)
@@ -64,6 +65,8 @@ export(rename_with_dt)
 export(replace_dt)
 export(replace_na_dt)
 export(right_join_dt)
+export(rleid)
+export(rleidv)
 export(rn_col)
 export(rowwise_dt)
 export(sample_dt)
