+++ vignettes/english_tutorial.Rmd
@@ -10,7 +10,7 @@ vignette: >
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
-  eval = FALSE,
+  #eval = FALSE,
   comment = "#>"
 )
 ```
@@ -113,7 +113,7 @@ relocate_dt(DF, V4) # reorder columns
 select_dt(DF, "V")
 select_dt(DF, "3$")
 select_dt(DF, ".2")
-select_dt(DF, "V1|X")
+select_dt(DF, "V1")
 select_dt(DF, -"^V2")
 # remove variables using "-" prior to function
 ```
@@ -146,101 +146,45 @@ summarise_dt(DF, uniqueN(V4))
 uniqueN(DF)
 ```
 
-## Add/update/delete columns
-```{r}
-### Modify a column
-DT[, V1 := V1^2]
-DT
-DF %>% set_mutate(V1 = V1^2)
-DF
-fsetequal(DT,DF)
-
-### Add one column
-DT[, v5 := log(V1)][] # adding [] prints the result
-set_mutate(DF, v5 = log(V1))
-fsetequal(DT,DF)
-
-### Add several columns
-DT[, c("v6", "v7") := .(sqrt(V1), "X")]
-DT[, ':='(v6 = sqrt(V1),
-          v7 = "X")]     # same, functional form
-
-set_mutate(DF, v6 = sqrt(V1), v7 = "X")
-
-fsetequal(DT,DF)
-
-### Create one column and remove the others
-DT[, .(v8 = V3 + 1)]
-transmute_dt(DF, v8 = V3 + 1)
-
-### Remove one column
-DT[, v5 := NULL]
-set_mutate(DF, v5 = NULL)
-fsetequal(DT,DF)
-
-### Remove several columns
-DT[, c("v6", "v7") := NULL]
-set_mutate(DF, v6 = NULL,v7 = NULL)
-fsetequal(DT,DF)
-
-### Remove columns using a vector of colnames
-cols <- c("V3")
-DT[, (cols) := NULL] # ! not DT[, cols := NULL]
-DF <- select_dt(DF, cols = cols,negate = TRUE)
-fsetequal(DT,DF)
-
-### Replace values for rows matching a condition
-DT[V2 < 4, V2 := 0L]
-DT
-DF = mutate_when(DF,V2 < 4, V2 = 0L)
-DF
-fsetequal(DT,DF)
-```
 
 ## group computation (by)
 ```{r}
 ### By group
+# not recommended
 DF %>%
   group_dt(
     by = V4,
     summarise_dt(sumV2 = sum(V2))
   )
+# recommended
+DF %>% 
+  summarise_dt(sumV2 = sum(V2),by = V4)
+
 
 ### By several groups
-DF %>%
-  group_dt(
-    by = .(V4,V1),
-    summarise_dt(sumV2 = sum(V2))
-  )
+DF %>% 
+  summarise_dt(sumV2 = sum(V2),by = .(V1,V4))
 
 ### Calling function in by
 DF %>% 
-  group_dt(
-    by = tolower(V4),
-    summarise_dt(sumV1 = sum(V1))
-  )
+  summarise_dt(sumV2 = sum(V2),by = tolower(V4))
+
 
 ### Assigning column name in by
 DF %>% 
-  group_dt(
-    by = .(abc = tolower(V4)),
-    summarise_dt(sumV1 = sum(V1))
-  )
+  summarise_dt(sumV2 = sum(V2),by = .(abc = tolower(V4)))
+
 
 ### Using a condition in by
 DF %>% 
-  group_dt(
-    by = V4 == "A",
-    summarise_dt(sum(V1))
-  )
[…]
