+++ vignettes/benchmark.Rmd
@@ -1,0 +1,176 @@
+---
+title: "Performance"
+output: rmarkdown::html_vignette
+author: Tian-Yuan Huang (huang.tian-yuan@qq.com)
+vignette: >
+  %\VignetteIndexEntry{Performance}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  warning = FALSE,
+  comment = "#>"
+)
+```
+  One may wonder how fast is *tidyfst*. Well, it depends. Generally, it is as fast as *data.table* because it is backed by it, but it would spend extra time on the generation of data.table codes. This extra time is marginal on large (and even small) data sets.<br><br>
+  Now let's do a test to compare the performance of *tidyfst*, *data.table* and *dplyr*. In the vignette we'll use a small data set. The example was provided by the *data.table* package (<https://h2oai.github.io/db-benchmark/>) and tweaked here. These tests are based on computation by groups.<br><br>
+  First let's load the package and generate some data.<br><br>
+```{r setup}
+# load packages
+library(tidyfst)
+library(data.table)
+library(dplyr)
+library(bench)
+
+# generate the data
+# if you have a HPC and want to try larger data sets, increase N
+N = 1e4 
+K = 1e2
+
+set.seed(2020)
+
+cat(sprintf("Producing data of %s rows and %s K groups factors\n", N, K))
+
+DT = data.table(
+
+  id1 = sample(sprintf("id%03d",1:K), N, TRUE),      # large groups (char)
+
+  id2 = sample(sprintf("id%03d",1:K), N, TRUE),      # large groups (char)
+
+  id3 = sample(sprintf("id%010d",1:(N/K)), N, TRUE), # small groups (char)
+
+  id4 = sample(K, N, TRUE),                          # large groups (int)
+
+  id5 = sample(K, N, TRUE),                          # large groups (int)
+
+  id6 = sample(N/K, N, TRUE),                        # small groups (int)
+
+  v1 =  sample(5, N, TRUE),                          # int in range [1,5]
+
+  v2 =  sample(5, N, TRUE),                          # int in range [1,5]
+
+  v3 =  round(runif(N,max=100),4)                    # numeric e.g. 23.5749
+
+)
+
+object_size(DT)
+```
+  This data is rather small, the size is around 527 Kb. However, with the *bench* package, we could detect the difference by increasing iteration times. In this way, examples listed here could be implemented even on relatively low performance computers. 
+  
+## Q1 
+  Here, we try to get median and standard deviation by groups.After dplyr v1.0.0, the regrouping feature could be confusing sometimes (comes with warning message). If you are using it, make sure they are in the right groups before grouped computation. In tidyfst and data.table, we have "by" parameter to specify the groups. Here we would not check if the results are equal, because dplyr will return a tibble class even when we input a data.table in the first place. The iteration time is 10 for each of the test below.
+```{r}
+bench::mark(
+  data.table = DT[,.(median_v3 = median(v3),
+                     sd_v3 = sd(v3)),
+                  by = .(id4,id5)],
+  tidyfst = DT %>%
+    summarise_dt(
+      by = "id4,id5",
+      median_v3 = median(v3),
+      sd_v3 = sd(v3)
+    ),
+  dplyr = DT %>%
+    group_by(id4,id5,.drop = TRUE) %>%
+    summarise(median_v3 = median(v3),sd_v3 = sd(v3)),
+  check = FALSE,iterations = 10
+) -> q1
+
+q1
[…]
