+++ man/reexports.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/_global_setting.R
+% Please edit documentation in R/_global_setting.R, R/global_setting.R
 \docType{import}
 \name{reexports}
 \alias{reexports}
@@ -21,6 +21,9 @@
 \alias{copy}
 \alias{key}
 \alias{CJ}
+\alias{rleid}
+\alias{rleidv}
+\alias{between}
 \title{Objects exported from other packages}
 \keyword{internal}
 \description{
@@ -28,8 +31,8 @@ These objects are imported from other packages. Follow the links
 below to see their documentation.
 
 \describe{
-  \item{data.table}{\code{\link[data.table]{as.data.table}}, \code{\link[data.table]{CJ}}, \code{\link[data.table]{copy}}, \code{\link[data.table]{data.table}}, \code{\link[data.table]{fcoalesce}}, \code{\link[data.table]{fintersect}}, \code{\link[data.table]{fread}}, \code{\link[data.table]{frollapply}}, \code{\link[data.table]{fsetdiff}}, \code{\link[data.table]{fsetequal}}, \code{\link[data.table]{funion}}, \code{\link[data.table]{fwrite}}, \code{\link[data.table]{key}}, \code{\link[data.table]{like}}, \code{\link[data.table]{rbindlist}}, \code{\link[data.table]{tables}}, \code{\link[data.table]{uniqueN}}}
+  \item{data.table}{\code{\link[data.table]{as.data.table}}, \code{\link[data.table]{as.data.table}}, \code{\link[data.table]{between}}, \code{\link[data.table]{CJ}}, \code{\link[data.table]{copy}}, \code{\link[data.table]{data.table}}, \code{\link[data.table]{data.table}}, \code{\link[data.table]{fcoalesce}}, \code{\link[data.table]{fcoalesce}}, \code{\link[data.table]{fintersect}}, \code{\link[data.table]{fintersect}}, \code{\link[data.table]{fread}}, \code{\link[data.table]{fread}}, \code{\link[data.table]{frollapply}}, \code{\link[data.table]{frollapply}}, \code{\link[data.table]{fsetdiff}}, \code{\link[data.table]{fsetdiff}}, \code{\link[data.table]{fsetequal}}, \code{\link[data.table]{fsetequal}}, \code{\link[data.table]{funion}}, \code{\link[data.table]{funion}}, \code{\link[data.table]{fwrite}}, \code{\link[data.table]{fwrite}}, \code{\link[data.table]{key}}, \code{\link[data.table]{like}}, \code{\link[data.table]{like}}, \code{\link[data.table]{rbindlist}}, \code{\link[data.table]{rbindlist}}, \code{\link[data.table]{rleid}}, \code{\link[data.table]{rleidv}}, \code{\link[data.table]{tables}}, \code{\link[data.table]{tables}}, \code{\link[data.table]{uniqueN}}}
 
-  \item{stringr}{\code{\link[stringr]{\%>\%}}}
+  \item{stringr}{\code{\link[stringr]{\%>\%}}, \code{\link[stringr]{\%>\%}}}
 }}
 
