+++ man/group_by.Rd
@@ -24,6 +24,9 @@ It can receive what \code{select_dt} receives.}
 A data.table with keys
 }
 \description{
+Carry out data manipulation within specified groups. Different from \code{group_dt},
+the implementation is split into two operations, namely grouping and implementation.
+
 Using \code{setkey} and \code{setkeyv} in \pkg{data.table}
 to carry out \code{group_by}-like functionalities in \pkg{dplyr}. This is
 not only convenient but also efficient in computation.

[…]
