+++ docs/JOSS/paper.bib
@@ -3,6 +3,7 @@
 DowleSrinivasan-480,
    Author = {Dowle, Matt and Srinivasan, Arun},
    Title = {data.table: Extension of `data.frame`},
+   Url = {https://CRAN.R-project.org/package=data.table},
    Note = {R package version 1.12.8},
    Year = {2019} }
 
@@ -13,6 +14,7 @@ DowleSrinivasan-480,
 WickhamFrancois-481,
    Author = {Wickham, Hadley and François, Romain and Henry, Lionel and Müller, Kirill},
    Title = {dplyr: A Grammar of Data Manipulation},
+   Url = {https://CRAN.R-project.org/package=dplyr},
    Note = {R package version 0.8.5},
    Year = {2020} }
 
@@ -22,7 +24,8 @@ WickhamFrancois-481,
 @misc{
 Henry-482,
    Author = {Henry, Lionel},
-   Title = {dtplyr: Data Table Back-End for 'dplyr'},
+   Title = {dtplyr: Data Table Back-End for 'dplyr'},   
+   Url = {https://CRAN.R-project.org/package=dtplyr},
    Note = {R package version 1.0.1},
    Year = {2020} }
 
@@ -32,7 +35,8 @@ Henry-482,
 @misc{
 Klik-483,
    Author = {Klik, Mark},
-   Title = {fst: Lightning Fast Serialization of Data Frames for R},
+   Title = {fst: Lightning Fast Serialization of Data Frames for R},   
+   Url = {https://CRAN.R-project.org/package=fst},
    Note = {R package version 0.9.2},
    Year = {2020} }
 
@@ -42,8 +46,8 @@ Klik-483,
 @misc{
 Demin-484,
    Author = {Demin, Gregory},
-   Title = {maditr: Fast Data Aggregation, Modification, and Filtering with Pipes
-and 'data.table'},
+   Title = {maditr: Fast Data Aggregation, Modification, and Filtering with Pipes and 'data.table'},    
+   Url = {https://CRAN.R-project.org/package=maditr},
    Note = {R package version 0.6.3},
    Year = {2019} }
 
@@ -51,20 +55,11 @@ and 'data.table'},
 
 
 @misc{
-RinkerKurkiewicz-485,
-   Author = {Rinker, Tyler and Kurkiewicz, Dason},
-   Title = {pacman: Package Management Tool},
-   Note = {R package version 0.5.1},
-   Year = {2019} }
-
-
-
-
-@misc{
 R-Core-Team-479,
-   Author = {R Core Team},
+   Author = {"R Core Team"},    
+   Url = {https://www.r-project.org/},
    Title = {R: A Language and Environment for Statistical Computing},
-   Year = {2020} }
+   Year = {2013} }
 
 
 
@@ -73,6 +68,7 @@ R-Core-Team-479,
 Sarda-Espinosa-486,
    Author = {Sarda-Espinosa, Alexis},
    Title = {table.express: Build 'data.table' Expressions with Data Manipulation Verbs},
+   Url = {https://CRAN.R-project.org/package=table.express},
    Note = {R package version 0.3.1},
    Year = {2019} }
 
@@ -87,8 +83,9 @@ Wickham-458,
    Volume = {59},
    Number = {10},
    Pages = {1-23},
-   Year = {2014} }
-
+   Year = {2014},
+   Url = {http://www.jstatsoft.org/v59/i10/},
+   Doi = {10.18637/jss.v059.i10} }
 
 
 
@@ -96,6 +93,7 @@ Wickham-458,
 Barrett-487,
    Author = {Barrett, Tyson},
    Title = {tidyfast: Fast Tidying of Data},
[…]
