### 2020-07-11 hope-data-science/tidyfst#4 @hope-data-science
Hello Robrecht Cannoodt!

Thank you for the comments. I have made more corrections and improvements on it, your suggestions have been very helpful. Below please find my responses to the suggestions.

1.  Please include a benchmark comparing dplyr, data.table and tidyfst using, for example, microbenchmark. You can include the benchmark either in the manuscript or in a separate vignette.

- Response: Done as suggested. You can find it in <https://hope-data-science.github.io/tidyfst/articles/benchmark.html#q1-1>.

2. Please write descriptions for each of the functions such that its functionality can be understood.

- Response: Done, would be released in v0.9.8.

3. Please write unit tests for each of the functions.

- Response: Done in two ways. (1) Automatically run and checked manually with *pkgdown* package and displayed in <https://hope-data-science.github.io/tidyfst/reference/index.html>, click into the functions and find the Example part. (2) Use a general tutorial to test out various functions and tasks, see <https://hope-data-science.github.io/tidyfst/articles/english_tutorial.html>.

4. Please include these sets of guidelines. usethis::use_tidy_github() sets up templates for each of these guidelines.

- Response: Done as suggested (e.g. <https://hope-data-science.github.io/tidyfst/CODE_OF_CONDUCT.html>).

5. Problems with the reference part.

- Response: Done as suggested. Find the new manuscript in <https://github.com/openjournals/joss-papers/blob/joss.02294/joss.02294/10.21105.joss.02294.pdf>.

Thank you for the contribution you have made to improve this work.

Best Regards,

Tian-Yuan Huang
