### 2020-07-11 #2388 @hope-data-science
@rcannood @mikldk Revision has been made and details could be found at <https://github.com/hope-data-science/tidyfst/issues/4>, thank you for the valuable comments. Let me know if there are more questions and suggestions.
