### 2020-07-14 #2388 @mikldk
@hope-data-science Thanks. I believe that @rcannood will now continue the review (note that many now has holiday which may also be true for @tomsing1).
