### 2020-07-15 hope-data-science/tidyfst#4 @rcannood
Hello Tian-Yuan Huang!

I will review your changes ASAP. At first glance everything looks good, except part 3.

Since tidyfst is meant to be used as a frequently used library for performing data manipulation, it is vital that each function is tested to verify that each function works as intended. You might be 100% that each of your functions work perfect right now, but as time goes on and R is updated, this might not always be the case.

Please add unit tests for each of your functions. Run `usethis::use_testthat()` to setup your project for unit testing. Run `usethis::use_test("arrange_dt.R")` to create a test file for `arrange_dt.R`. See [http://r-pkgs.had.co.nz/tests.html](http://r-pkgs.had.co.nz/tests.html) for more information on how to write unit tests. You can also have a look at dplyr to see how unit tests are performed in this package: [https://github.com/tidyverse/dplyr/tree/master/tests/testthat](https://github.com/tidyverse/dplyr/tree/master/tests/testthat).

Kind regards,
Robrecht
