### 2020-07-22 ff657bb
Update mutate_dt.R
For <https://github.com/hope-data-science/tidyfst/issues/6>. Done.
