### 2020-07-27 #2388 @mikldk
@hope-data-science I apologise for the delay, but am confident that both @tomsing1 and @rcannood will look at this when they are back from holiday. @tomsing1, @rcannood: Can you give a rough estimate of when you will be able to finish this review?
