### 2020-07-27 #2388 @hope-data-science
> @hope-data-science I apologise for the delay, but am confident that both @tomsing1 and @rcannood will look at this when they are back from holiday. @tomsing1, @rcannood: Can you give a rough estimate of when you will be able to finish this review?

Thank you for the feedback. I am currently working on the unit test of the package (from comments of @rcannood ), which will take weeks as some functions have some details to test out. I'll make feedbacks when there are new progresses.

Thanks.
