### 2020-07-29 e0100cb
Delete global_setting.R
