+++ R/global_setting.R
@@ -1,51 +1,0 @@
-#' @import data.table
-#' @import fst
-#' @import stringr
-#' @importFrom stats na.omit setNames
-
-
-#' @export
-stringr::`%>%`
-
-#' @export
-data.table::data.table
-
-#' @export
-data.table::as.data.table
-
-#' @export
-data.table::fread
-
-#' @export
-data.table::fwrite
-
-#' @export
-data.table::fintersect
-
-#' @export
-data.table::fsetdiff
-
-#' @export
-data.table::funion
-
-#' @export
-data.table::fsetequal
-
-#' @export
-data.table::rbindlist
-
-#' @export
-data.table::frollapply
-
-#' @export
-data.table::between
-
-#' @export
-data.table::fcoalesce
-
-#' @export
-data.table::tables
-
-#' @export
-data.table::like
-
