+++ NAMESPACE
@@ -8,7 +8,6 @@ export(arrange_dt)
 export(as.data.table)
 export(as_dt)
 export(as_fst)
-export(between)
 export(chop_dt)
 export(col_rn)
 export(complete_dt)
