+++ R/slice.R
@@ -137,106 +137,132 @@ slice_tail_dt = function(.data,n,by = NULL){
 
 #' @rdname slice
 #' @export
-slice_max_dt = function(.data,order_by,n,with_ties = TRUE,by = NULL){
-  .data = as_dt(.data)
-  order_by = deparse(substitute(order_by))
-  by_char = deparse(substitute(by))
+slice_max_dt = function(.data,order_by,n,by = NULL,with_ties = TRUE){
+  dt = as_dt(.data)
+  wt = substitute(order_by) %>% deparse()
+  by = substitute(by) %>% deparse()
+  setorderv(dt,cols = wt,order = -1L)
+  if (n > 0 & n < 1)
+    n = nrow(.data) * n
+  else if (n <= 0)
+    stop("Invalid input, n should take a positive value.")
 
-  if(by_char == "NULL"){
-    if(n > 0 & n < 1) n = nrow(.data) * n
-    else if(n <= 0) stop("Invalid input, n should take a positive value.")
-    tm_ = ifelse(with_ties,"min","first")
-    index = frankv(.data[[order_by]],order = -1,ties.method = tm_) <= n
-    .data[index]
-  }else{
-    eval(substitute(
-      .data[,{
-        if (n > 0 & n < 1)
-          n = nrow(.SD) * n
-        else if (n <= 0)
-          stop("Invalid input, n should take a positive value.")
-        tm_ = ifelse(with_ties, "min", "first")
-        index = frankv(.SD[[order_by]], order = -1, ties.method = tm_) <= n
-        .SD[index]
-      },by]
-    ))
-  }
-}
-
-# slice_max_dt = function(.data,order_by,n,with_ties = TRUE){
-#   .data = as_dt(.data)
-#   if(n > 0 & n < 1) n = nrow(.data) * n
-#   else if(n <= 0) stop("Invalid input, n should take a positive value.")
-#
-#   order_by = deparse(substitute(order_by))
-#   tm_ = ifelse(with_ties,"min","first")
-#   index = frankv(.data[[order_by]],order = -1,ties.method = tm_) <= n
-#
-#   .data[index]
-# }
-
-#' @rdname slice
-#' @export
-slice_min_dt = function(.data,order_by,n,with_ties = TRUE,by = NULL){
-  .data = as_dt(.data)
-  order_by = deparse(substitute(order_by))
-  by_char = deparse(substitute(by))
-
-  if(by_char == "NULL"){
-    if(n > 0 & n < 1) n = nrow(.data) * n
-    else if(n <= 0) stop("Invalid input, n should take a positive value.")
-    tm_ = ifelse(with_ties,"min","first")
-    index = frankv(.data[[order_by]],order = 1,ties.method = tm_) <= n
-    .data[index]
-  }else{
-    eval(substitute({
-      .data[,{
-        if (n > 0 & n < 1)
-          n = nrow(.SD) * n
-        else if (n <= 0)
-          stop("Invalid input, n should take a positive value.")
-        tm_ = ifelse(with_ties, "min", "first")
-        index = frankv(.SD[[order_by]], order = 1, ties.method = tm_) <= n
-        .SD[index]
-      },by]
-    }))
-  }
-}
-
-# slice_min_dt = function(.data,order_by,n,with_ties = TRUE){
-#   .data = as_dt(.data)
-#   if(n > 0 & n < 1) n = nrow(.data) * n
-#   else if(n <= 0) stop("Invalid input, n should take a positive value.")
-#
-#   order_by = deparse(substitute(order_by))
-#   tm_ = ifelse(with_ties,"min","first")
-#   index = frankv(.data[[order_by]],order = 1,ties.method = tm_) <= n
-#
-#   .data[index]
-# }
-
-
-#' @rdname slice
-#' @export
[…]
