+++ R/mutate_dt.R
@@ -1,3 +1,4 @@
+
 #' @title Mutate columns in data.frame
 #' @description Adds or updates columns in data.frame.
 #' @param .data data.frame
@@ -26,7 +27,7 @@ mutate_dt = function(.data,...,by){
 
 transmute_dt = function(.data,...,by){
   dt = as_dt(.data)
-  substitute(dt[,.(id_ = 1:.N,...),by][]) %>% eval() %>% 
+  substitute(dt[,.(id_ = 1:.N,...),by][]) %>% eval() %>%
     .[,id_:=NULL] %>% .[]
 }
 
@@ -35,4 +36,6 @@ transmute_dt = function(.data,...,by){
 #   dt = as_dt(.data)
 #   substitute(dt[,.(...),by][]) %>% eval()
 # }
+
+
 
