### 2020-07-31 hope-data-science/tidyfst#4 @hope-data-science
Hello Robrecht!

The unit tests have been carried out in <https://github.com/hope-data-science/tidyfst/tree/master/tests>. Thank you for the suggestions. The software has improved in this process. If there are further questions, feel free to let me know.

Thanks.

Best,
Tian-Yuan Huang
