+++ R/on_attach.R
@@ -1,7 +1,7 @@
 .onAttach = function(...) {
     options(
         "datatable.print.class" = TRUE, # print class in data.table
-        #    "datatable.print.trunc.cols" = TRUE,
+           "datatable.print.trunc.cols" = TRUE,
            "datatable.print.keys" = TRUE
         )
     hints = c(
