### 2020-08-05 #2388 @hope-data-science
The unit tests have been completed, see updates in <https://github.com/hope-data-science/tidyfst/issues/4>.

Thanks.
