### 2020-08-07 hope-data-science/tidyfst#4 @rcannood
Hello Tian-Yuan,

Thank you for your hard work, all my previous comments have been addressed. And sorry for the late response, I was on holidays.

I forked the tidyfst repository in order to automate the R CMD check. I notice that the R CMD check generates a lot of warnings:

  <summary>R CMD check output</summary>

```
==> devtools::check()

Updating tidyfst documentation
Updating roxygen version in /home/rcannood/workspace_vib/tidyfst/DESCRIPTION
Loading tidyfst

Life's short, use R.

Writing NAMESPACE
Writing NAMESPACE
Writing reexports.Rd
── Building ──────────────────────────────────────────────────────────────────── tidyfst ──
Setting env vars:
Deleting all_at_if.Rd
Deleting longer_dt.Rd
Deleting mutate_when.Rd
Deleting select_dt.Rd
Deleting set_dt.Rd
Deleting set.Rd
Deleting slice_dt.Rd
● CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
● CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
● CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
───────────────────────────────────────────────────────────────────────────────────────────
✓  checking for file ‘/home/rcannood/workspace_vib/tidyfst/DESCRIPTION’ ...
─  preparing ‘tidyfst’:
✓  checking DESCRIPTION meta-information ...
✓  creating vignettes (12.6s)
─  excluding invalid files
   Subdirectory 'R' contains invalid file names:
     ‘_global_setting.R’
─  building ‘tidyfst_0.9.8.tar.gz’

── Checking ──────────────────────────────────────────────────────────────────── tidyfst ──
Setting env vars:
● _R_CHECK_CRAN_INCOMING_USE_ASPELL_: TRUE
● _R_CHECK_CRAN_INCOMING_REMOTE_    : FALSE
● _R_CHECK_CRAN_INCOMING_           : FALSE
● _R_CHECK_FORCE_SUGGESTS_          : FALSE
● NOT_CRAN                          : true
── R CMD check ─────────────────────────────────────────────────────────────────
─  using log directory ‘/home/rcannood/workspace_vib/tidyfst.Rcheck’
─  using R version 4.0.2 (2020-06-22)
─  using platform: x86_64-pc-linux-gnu (64-bit)
─  using session charset: UTF-8
─  using options ‘--no-manual --as-cran’
✓  checking for file ‘tidyfst/DESCRIPTION’
─  this is package ‘tidyfst’ version ‘0.9.8’
─  package encoding: UTF-8
✓  checking package namespace information ...
✓  checking package dependencies (1.9s)
✓  checking if this is a source package ...
✓  checking if there is a namespace
✓  checking for executable files (426ms)
✓  checking for hidden files and directories
W  checking for portable file names ...
   Found the following file with a non-portable file name:
     man/figures/hex-tidyfst .png
   These are not fully portable file names.
   See section ‘Package structure’ in the ‘Writing R Extensions’ manual.
✓  checking for sufficient/correct file permissions ...
✓  checking serialization versions ...
✓  checking whether package ‘tidyfst’ can be installed (1.7s)
✓  checking installed package size ...
✓  checking package directory ...
✓  checking for future file timestamps (558ms)
✓  checking ‘build’ directory
✓  checking DESCRIPTION meta-information ...
N  checking top-level files
   Non-standard file/directory found at top level:
     ‘cran-comments.md’
✓  checking for left-over files
✓  checking index information ...
✓  checking package subdirectories ...
✓  checking R files for non-ASCII characters ...
✓  checking R files for syntax errors ...
✓  checking whether the package can be loaded ...
✓  checking whether the package can be loaded with stated dependencies ...
✓  checking whether the package can be unloaded cleanly ...
✓  checking whether the namespace can be loaded with stated dependencies ...
✓  checking whether the namespace can be unloaded cleanly ...
✓  checking loading without being on the library search path ...
✓  checking dependencies in R code ...
✓  checking S3 generic/method consistency (495ms)
✓  checking replacement functions ...
✓  checking foreign function calls ...
✓  checking R code for possible problems (2.4s)
✓  checking Rd files ...
✓  checking Rd metadata ...
✓  checking Rd line widths ...
N  checking Rd cross-references (1.6s)
   Package unavailable to check Rd xrefs: ‘fastDummies’
✓  checking for missing documentation entries ...
✓  checking for code/documentation mismatches (848ms)
W  checking Rd \usage sections ...
   Undocumented arguments in documentation object 'rename_dt'
     ‘.fn’

   Functions with \usage entries need to have the appropriate \alias
   entries, and all their arguments documented.
   The \usage entries must correspond to syntactically valid R code.
   See chapter ‘Writing R documentation files’ in the ‘Writing R
   Extensions’ manual.
✓  checking Rd contents (685ms)
✓  checking for unstated dependencies in examples ...
✓  checking installed files from ‘inst/doc’ ...
✓  checking files in ‘vignettes’ ...
E  checking examples (1.8s)
   Running examples in ‘tidyfst-Ex.R’ failed
   The error most likely occurred in:

   > base::assign(".ptime", proc.time(), pos = "CheckExEnv")
   > ### Name: slice_dt
   > ### Title: Subset rows using their positions
   > ### Aliases: slice_dt slice_head_dt slice_tail_dt slice_max_dt slice_min_dt
   > ###   slice_sample_dt
   >
   > ### ** Examples
   >
   >
   > a = iris
   > slice_dt(a,1,2)
      Sepal.Length Sepal.Width Petal.Length Petal.Width Species
             <num>       <num>        <num>       <num>  <fctr>
   1:          5.1         3.5          1.4         0.2  setosa
   2:          4.9         3.0          1.4         0.2  setosa
   > slice_dt(a,2:3)
      Sepal.Length Sepal.Width Petal.Length Petal.Width Species
             <num>       <num>        <num>       <num>  <fctr>
   1:          4.9         3.0          1.4         0.2  setosa
   2:          4.7         3.2          1.3         0.2  setosa
   > slice_dt(a,141:.N)
       Sepal.Length Sepal.Width Petal.Length Petal.Width   Species
              <num>       <num>        <num>       <num>    <fctr>
    1:          6.7         3.1          5.6         2.4 virginica
    2:          6.9         3.1          5.1         2.3 virginica
    3:          5.8         2.7          5.1         1.9 virginica
    4:          6.8         3.2          5.9         2.3 virginica
    5:          6.7         3.3          5.7         2.5 virginica
    6:          6.7         3.0          5.2         2.3 virginica
    7:          6.3         2.5          5.0         1.9 virginica
    8:          6.5         3.0          5.2         2.0 virginica
    9:          6.2         3.4          5.4         2.3 virginica
   10:          5.9         3.0          5.1         1.8 virginica
   > slice_dt(a,1,.N)
      Sepal.Length Sepal.Width Petal.Length Petal.Width   Species
             <num>       <num>        <num>       <num>    <fctr>
   1:          5.1         3.5          1.4         0.2    setosa
   2:          5.9         3.0          5.1         1.8 virginica
   > slice_head_dt(a,5)
      Sepal.Length Sepal.Width Petal.Length Petal.Width Species
             <num>       <num>        <num>       <num>  <fctr>
   1:          5.1         3.5          1.4         0.2  setosa
   2:          4.9         3.0          1.4         0.2  setosa
   3:          4.7         3.2          1.3         0.2  setosa
   4:          4.6         3.1          1.5         0.2  setosa
   5:          5.0         3.6          1.4         0.2  setosa
   > slice_head_dt(a,0.1)
       Sepal.Length Sepal.Width Petal.Length Petal.Width Species
              <num>       <num>        <num>       <num>  <fctr>
    1:          5.1         3.5          1.4         0.2  setosa
    2:          4.9         3.0          1.4         0.2  setosa
    3:          4.7         3.2          1.3         0.2  setosa
    4:          4.6         3.1          1.5         0.2  setosa
    5:          5.0         3.6          1.4         0.2  setosa
    6:          5.4         3.9          1.7         0.4  setosa
    7:          4.6         3.4          1.4         0.3  setosa
    8:          5.0         3.4          1.5         0.2  setosa
    9:          4.4         2.9          1.4         0.2  setosa
   10:          4.9         3.1          1.5         0.1  setosa
   11:          5.4         3.7          1.5         0.2  setosa
   12:          4.8         3.4          1.6         0.2  setosa
   13:          4.8         3.0          1.4         0.1  setosa
   14:          4.3         3.0          1.1         0.1  setosa
   15:          5.8         4.0          1.2         0.2  setosa
   > slice_tail_dt(a,5)
      Sepal.Length Sepal.Width Petal.Length Petal.Width   Species
             <num>       <num>        <num>       <num>    <fctr>
   1:          6.7         3.0          5.2         2.3 virginica
   2:          6.3         2.5          5.0         1.9 virginica
   3:          6.5         3.0          5.2         2.0 virginica
   4:          6.2         3.4          5.4         2.3 virginica
   5:          5.9         3.0          5.1         1.8 virginica
   > slice_tail_dt(a,0.1)
       Sepal.Length Sepal.Width Petal.Length Petal.Width   Species
              <num>       <num>        <num>       <num>    <fctr>
    1:          7.7         3.0          6.1         2.3 virginica
    2:          6.3         3.4          5.6         2.4 virginica
    3:          6.4         3.1          5.5         1.8 virginica
    4:          6.0         3.0          4.8         1.8 virginica
    5:          6.9         3.1          5.4         2.1 virginica
    6:          6.7         3.1          5.6         2.4 virginica
    7:          6.9         3.1          5.1         2.3 virginica
    8:          5.8         2.7          5.1         1.9 virginica
    9:          6.8         3.2          5.9         2.3 virginica
   10:          6.7         3.3          5.7         2.5 virginica
   11:          6.7         3.0          5.2         2.3 virginica
   12:          6.3         2.5          5.0         1.9 virginica
   13:          6.5         3.0          5.2         2.0 virginica
   14:          6.2         3.4          5.4         2.3 virginica
   15:          5.9         3.0          5.1         1.8 virginica
   > slice_max_dt(a,Sepal.Length,10)
       Sepal.Length Sepal.Width Petal.Length Petal.Width   Species
              <num>       <num>        <num>       <num>    <fctr>
    1:          7.9         3.8          6.4         2.0 virginica
    2:          7.7         3.8          6.7         2.2 virginica
    3:          7.7         2.6          6.9         2.3 virginica
    4:          7.7         2.8          6.7         2.0 virginica
    5:          7.7         3.0          6.1         2.3 virginica
    6:          7.6         3.0          6.6         2.1 virginica
    7:          7.4         2.8          6.1         1.9 virginica
    8:          7.3         2.9          6.3         1.8 virginica
    9:          7.2         3.6          6.1         2.5 virginica
   10:          7.2         3.2          6.0         1.8 virginica
   11:          7.2         3.0          5.8         1.6 virginica
   > slice_max_dt(a,Sepal.Length,10,with_ties = FALSE)
       Sepal.Length Sepal.Width Petal.Length Petal.Width   Species
              <num>       <num>        <num>       <num>    <fctr>
    1:          7.9         3.8          6.4         2.0 virginica
    2:          7.7         3.8          6.7         2.2 virginica
    3:          7.7         2.6          6.9         2.3 virginica
    4:          7.7         2.8          6.7         2.0 virginica
    5:          7.7         3.0          6.1         2.3 virginica
    6:          7.6         3.0          6.6         2.1 virginica
    7:          7.4         2.8          6.1         1.9 virginica
    8:          7.3         2.9          6.3         1.8 virginica
    9:          7.2         3.6          6.1         2.5 virginica
   10:          7.2         3.2          6.0         1.8 virginica
   > slice_min_dt(a,Sepal.Length,10)
       Sepal.Length Sepal.Width Petal.Length Petal.Width Species
              <num>       <num>        <num>       <num>  <fctr>
    1:          4.3         3.0          1.1         0.1  setosa
    2:          4.4         2.9          1.4         0.2  setosa
    3:          4.4         3.0          1.3         0.2  setosa
    4:          4.4         3.2          1.3         0.2  setosa
    5:          4.5         2.3          1.3         0.3  setosa
    6:          4.6         3.1          1.5         0.2  setosa
    7:          4.6         3.4          1.4         0.3  setosa
    8:          4.6         3.6          1.0         0.2  setosa
    9:          4.6         3.2          1.4         0.2  setosa
   10:          4.7         3.2          1.3         0.2  setosa
   11:          4.7         3.2          1.6         0.2  setosa
   > slice_min_dt(a,Sepal.Length,10,with_ties = FALSE)
       Sepal.Length Sepal.Width Petal.Length Petal.Width Species
              <num>       <num>        <num>       <num>  <fctr>
    1:          4.3         3.0          1.1         0.1  setosa
    2:          4.4         2.9          1.4         0.2  setosa
    3:          4.4         3.0          1.3         0.2  setosa
    4:          4.4         3.2          1.3         0.2  setosa
    5:          4.5         2.3          1.3         0.3  setosa
    6:          4.6         3.1          1.5         0.2  setosa
    7:          4.6         3.4          1.4         0.3  setosa
    8:          4.6         3.6          1.0         0.2  setosa
    9:          4.6         3.2          1.4         0.2  setosa
   10:          4.7         3.2          1.3         0.2  setosa
   > slice_sample_dt(a,10)
       Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
              <num>       <num>        <num>       <num>     <fctr>
    1:          5.8         2.7          4.1         1.0 versicolor
    2:          6.4         2.8          5.6         2.1  virginica
    3:          4.4         3.2          1.3         0.2     setosa
    4:          4.3         3.0          1.1         0.1     setosa
    5:          7.0         3.2          4.7         1.4 versicolor
    6:          5.4         3.0          4.5         1.5 versicolor
    7:          5.4         3.4          1.7         0.2     setosa
    8:          7.6         3.0          6.6         2.1  virginica
    9:          6.1         2.8          4.7         1.2 versicolor
   10:          4.6         3.4          1.4         0.3     setosa
   > slice_sample_dt(a,0.1)
       Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
              <num>       <num>        <num>       <num>     <fctr>
    1:          6.3         2.5          4.9         1.5 versicolor
    2:          6.0         2.9          4.5         1.5 versicolor
    3:          5.5         3.5          1.3         0.2     setosa
    4:          6.5         3.0          5.8         2.2  virginica
    5:          7.2         3.6          6.1         2.5  virginica
    6:          5.5         4.2          1.4         0.2     setosa
    7:          7.6         3.0          6.6         2.1  virginica
    8:          7.2         3.2          6.0         1.8  virginica
    9:          5.6         3.0          4.1         1.3 versicolor
   10:          5.2         4.1          1.5         0.1     setosa
   11:          6.0         2.7          5.1         1.6 versicolor
   12:          5.6         2.5          3.9         1.1 versicolor
   13:          6.1         2.8          4.7         1.2 versicolor
   14:          4.5         2.3          1.3         0.3     setosa
   15:          6.5         3.2          5.1         2.0  virginica
   >
   >
   > # use by to slice by group
   >
   > ## following codes get the same results
   > slice_dt(a,1:3,by = "Species")
         Species Sepal.Length Sepal.Width Petal.Length Petal.Width
          <fctr>        <num>       <num>        <num>       <num>
   1:     setosa          5.1         3.5          1.4         0.2
   2:     setosa          4.9         3.0          1.4         0.2
   3:     setosa          4.7         3.2          1.3         0.2
   4: versicolor          7.0         3.2          4.7         1.4
   5: versicolor          6.4         3.2          4.5         1.5
   6: versicolor          6.9         3.1          4.9         1.5
   7:  virginica          6.3         3.3          6.0         2.5
   8:  virginica          5.8         2.7          5.1         1.9
   9:  virginica          7.1         3.0          5.9         2.1
   > slice_dt(a,1:3,by = Species)
         Species Sepal.Length Sepal.Width Petal.Length Petal.Width
          <fctr>        <num>       <num>        <num>       <num>
   1:     setosa          5.1         3.5          1.4         0.2
   2:     setosa          4.9         3.0          1.4         0.2
   3:     setosa          4.7         3.2          1.3         0.2
   4: versicolor          7.0         3.2          4.7         1.4
   5: versicolor          6.4         3.2          4.5         1.5
   6: versicolor          6.9         3.1          4.9         1.5
   7:  virginica          6.3         3.3          6.0         2.5
   8:  virginica          5.8         2.7          5.1         1.9
   9:  virginica          7.1         3.0          5.9         2.1
   > slice_dt(a,1:3,by = .(Species))
         Species Sepal.Length Sepal.Width Petal.Length Petal.Width
          <fctr>        <num>       <num>        <num>       <num>
   1:     setosa          5.1         3.5          1.4         0.2
   2:     setosa          4.9         3.0          1.4         0.2
   3:     setosa          4.7         3.2          1.3         0.2
   4: versicolor          7.0         3.2          4.7         1.4
   5: versicolor          6.4         3.2          4.5         1.5
   6: versicolor          6.9         3.1          4.9         1.5
   7:  virginica          6.3         3.3          6.0         2.5
   8:  virginica          5.8         2.7          5.1         1.9
   9:  virginica          7.1         3.0          5.9         2.1
   >
   > slice_head_dt(a,2,by = Species)
         Species Sepal.Length Sepal.Width Petal.Length Petal.Width
          <fctr>        <num>       <num>        <num>       <num>
   1:     setosa          5.1         3.5          1.4         0.2
   2:     setosa          4.9         3.0          1.4         0.2
   3: versicolor          7.0         3.2          4.7         1.4
   4: versicolor          6.4         3.2          4.5         1.5
   5:  virginica          6.3         3.3          6.0         2.5
   6:  virginica          5.8         2.7          5.1         1.9
   > slice_tail_dt(a,2,by = Species)
         Species Sepal.Length Sepal.Width Petal.Length Petal.Width
          <fctr>        <num>       <num>        <num>       <num>
   1:     setosa          5.3         3.7          1.5         0.2
   2:     setosa          5.0         3.3          1.4         0.2
   3: versicolor          5.1         2.5          3.0         1.1
   4: versicolor          5.7         2.8          4.1         1.3
   5:  virginica          6.2         3.4          5.4         2.3
   6:  virginica          5.9         3.0          5.1         1.8
   >
   > slice_max_dt(a,Sepal.Length,3,by = Species)
          Species Sepal.Length Sepal.Width Petal.Length Petal.Width
           <fctr>        <num>       <num>        <num>       <num>
    1:  virginica          7.9         3.8          6.4         2.0
    2:  virginica          7.7         3.8          6.7         2.2
    3:  virginica          7.7         2.6          6.9         2.3
    4:  virginica          7.7         2.8          6.7         2.0
    5:  virginica          7.7         3.0          6.1         2.3
    6: versicolor          7.0         3.2          4.7         1.4
    7: versicolor          6.9         3.1          4.9         1.5
    8: versicolor          6.8         2.8          4.8         1.4
    9:     setosa          5.8         4.0          1.2         0.2
   10:     setosa          5.7         4.4          1.5         0.4
[…]
