### 2020-08-07 hope-data-science/tidyfst#4 @rcannood
If you want, you can merge the commits I made in my repository. If you do, Github Actions will automatically run the R CMD check at every commit and tell you if something is wrong. The UI looks like this:

![Screenshot from 2020-08-07 11-23-49](https://user-images.githubusercontent.com/553642/89631040-7f47cb80-d8a0-11ea-8453-c2c1ade6b621.png)
