+++ NAMESPACE
@@ -22,6 +22,7 @@ export(distinct_dt)
 export(drop_na_dt)
 export(dummy_dt)
 export(export_fst)
+export(fcase)
 export(fcoalesce)
 export(fill_na_dt)
 export(filter_dt)
