+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: tidyfst
 Title: Tidy Verbs for Fast Data Manipulation
-Version: 0.9.8
+Version: 0.9.9
 Depends: R (>= 3.3.0)
 Authors@R: 
     person(given = "Tian-Yuan",
@@ -20,7 +20,7 @@ Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.1.0
 Imports: 
-    data.table (>= 1.12.8),
+    data.table (>= 1.13.0),
     fst (>= 0.9.0),
     stringr (>= 1.4.0)
 Suggests: 
