+++ R/rename_dt.R
@@ -4,6 +4,8 @@
 #' @param .data data.frame
 #' @param ... statements of rename, e.g. `sl = Sepal.Length` means the column named
 #' as "Sepal.Length" would be renamed to "sl"
+#' @param .fn A function used to transform the selected columns.
+#'  Should return a character vector the same length as the input.
 #' @return data.table
 #' @seealso \code{\link[dplyr]{rename}}
 #' @examples
