### 2020-08-07 hope-data-science/tidyfst#4 @hope-data-science
Hello Robrecht!

Thank you so much for the valuable comments you have contributed. The previous upload was a bit early, I've resubmitted the package to CRAN which returns similar results as in the check you have provided, and I've fixed them and resubmit it both to CRAN and GitHub. If there are further issues, please feel free to let me know.

Thanks.
