### 2020-08-07 hope-data-science/tidyfst#4 @rcannood
That's good to hear, that explains!

I still see duplicated files in the repository, for example, man/longer.Rd and man/longer_dt.Rd.

When you remove the duplicate files, go ahead and close this issue.
