### 2020-08-07 ab70b5f
Delete longer_dt.Rd
