### 2020-08-07 #2388 @rcannood
The changes made by @hope-data-science address all of my comments.

@hope-data-science, could you ask whedon to regenerate the pdf?
