### 2020-08-07 2c273d2
Delete slice_dt.Rd
