### 2020-08-07 615a04f
Delete all_at_if.Rd
