### 2020-08-07 hope-data-science/tidyfst#4 @hope-data-science
I've deleted the duplicated and deprecated ones. *tidyfst* would not improve so much without your help.  If there are further questions, let me know~

Thanks.
