+++ docs/JOSS/paper.bib
@@ -1,4 +1,4 @@
-﻿
+
 @misc{
 DowleSrinivasan-480,
    Author = {Dowle, Matt and Srinivasan, Arun},
@@ -24,7 +24,7 @@ WickhamFrancois-481,
 @misc{
 Henry-482,
    Author = {Henry, Lionel},
-   Title = {dtplyr: Data Table Back-End for 'dplyr'},   
+   Title = {dtplyr: Data Table Back-End for 'dplyr'},
    Url = {https://CRAN.R-project.org/package=dtplyr},
    Note = {R package version 1.0.1},
    Year = {2020} }
@@ -35,7 +35,7 @@ Henry-482,
 @misc{
 Klik-483,
    Author = {Klik, Mark},
-   Title = {fst: Lightning Fast Serialization of Data Frames for R},   
+   Title = {fst: Lightning Fast Serialization of Data Frames for R},
    Url = {https://CRAN.R-project.org/package=fst},
    Note = {R package version 0.9.2},
    Year = {2020} }
@@ -46,7 +46,7 @@ Klik-483,
 @misc{
 Demin-484,
    Author = {Demin, Gregory},
-   Title = {maditr: Fast Data Aggregation, Modification, and Filtering with Pipes and 'data.table'},    
+   Title = {maditr: Fast Data Aggregation, Modification, and Filtering with Pipes and 'data.table'},
    Url = {https://CRAN.R-project.org/package=maditr},
    Note = {R package version 0.6.3},
    Year = {2019} }
@@ -56,10 +56,13 @@ Demin-484,
 
 @misc{
 R-Core-Team-479,
-   Author = {"R Core Team"},    
-   Url = {https://www.r-project.org/},
-   Title = {R: A Language and Environment for Statistical Computing},
-   Year = {2013} }
+  title = {R: A Language and Environment for Statistical Computing},
+  author = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address = {Vienna, Austria},
+  year = {2020},
+  url = {https://www.R-project.org/}
+}
 
 
 
@@ -103,7 +106,7 @@ Barrett-487,
 @misc{
 Huang-488,
    Author = {Huang, Tian-Yuan},
-   Title = {tidyfst: Tidy Verbs for Fast Data Manipulation},   
+   Title = {tidyfst: Tidy Verbs for Fast Data Manipulation},
    Url = {https://CRAN.R-project.org/package=tidyfst},
    Note = {R package version 0.9.7},
    Year = {2020} }
@@ -114,7 +117,7 @@ Huang-488,
 @misc{
 Huang-489,
    Author = {Huang, Tian-Yuan},
-   Title = {tidyft: Tidy Verbs for Fast Data Operations by Reference},  
+   Title = {tidyft: Tidy Verbs for Fast Data Operations by Reference},
    Url = {https://CRAN.R-project.org/package=tidyft},
    Note = {R package version 0.4.5},
    Year = {2020} }
@@ -125,7 +128,7 @@ Huang-489,
 @misc{
 WickhamHenry-490,
    Author = {Wickham, Hadley and Henry, Lionel},
-   Title = {tidyr: Tidy Messy Data},   
+   Title = {tidyr: Tidy Messy Data},
    Url = {https://CRAN.R-project.org/package=tidyr},
    Note = {R package version 1.0.2},
    Year = {2020} }
@@ -136,7 +139,7 @@ WickhamHenry-490,
 @misc{
 Fairbanks-491,
    Author = {Fairbanks, Mark},
-   Title = {tidytable: Tidy Interface to 'data.table'},  
+   Title = {tidytable: Tidy Interface to 'data.table'},
[…]
