### 2020-08-08 #2388 @hope-data-science
Make new revision on the reference part. @whedon generate pdf
