### 2020-08-11 hope-data-science/tidyfst#7 @tomsing1
[JOSS Review] Non-portable filenames
Thanks for submitting your paper to JOSS and for addressing so many of the comments already!

I just ran R CMD check and got the following error, related to filenames in your repo:

```
> checking for portable file names ... ERROR
  Found the following file with a non-portable file name:
    docs/R语言数据操纵之美_黄天元.pptx
  These are not valid file names on all R platforms.
  Please rename the files and try again.
  See section ‘Package structure’ in the ‘Writing R Extensions’ manual.
   WARNING
  Found the following files with non-portable file names:
    docs/R语言数据操纵之美_黄天元.pptx
    docs/reference/figures/hex-tidyfst .png
    man/figures/hex-tidyfst .png
  These are not fully portable file names.
  See section ‘Package structure’ in the ‘Writing R Extensions’ manual.
```

Perhaps you want to rename the three files (the last two have spurious spaces before the suffix)?
- docs/R语言数据操纵之美_黄天元.pptx
- docs/reference/figures/hex-tidyfst .png
- man/figures/hex-tidyfst .png

Best,
Thomas
