### 2020-08-11 #2388 @mikldk
When hope-data-science/tidyfst#7 is closed we should be close to finishing the review.
