### 2020-08-11 hope-data-science/tidyfst#7 @hope-data-science
Thank you for the feedback. I've removed the figures which could still work for the current logo. And the file with filename of Chinese has been removed (have changed to an English name and should work later). Hope this time it could work well.

Thanks.
