### 2020-08-11 hope-data-science/tidyfst#7 @tomsing1
Great - thanks for addressing this so quickly!
