### 2020-08-13 #2388 @mikldk
Comments:

* A missing space before reference here: "supported by fst package(Klik, 2020)."
* When I go to <https://github.com/hope-data-science/tidyfst/graphs/contributors> I also only see @hope-data-science (Tian-Yuan Huang) contributing. Also, in the paper under "Acknowledgement" a singular "me" is used - yet the paper has two authors.  Can you please here in this review thread briefly describe the contributions of Bin Zhao?
