### 2020-08-13 #2388 @hope-data-science
Response to comments:

•A missing space before reference here: "supported by fst package(Klik, 2020)."
**Reply**: Corrected as suggested.

•When I go to https://github.com/hope-data-science/tidyfst/graphs/contributors I also only see @hope-data-science (Tian-Yuan Huang) contributing. Also, in the paper under "Acknowledgement" a singular "me" is used - yet the paper has two authors. Can you please here in this review thread briefly describe the contributions of Bin Zhao?
**Reply**: The "Acknowledgement" part was drafted long ago (when I start tidyfst as a small project of my own) and keep the same as in the GitHub README, I changed it to us in the manuscript. I am currently a PhD candidate in Fudan University and Bin Zhao is my supervisor. At later stage of the software development, I've used tidyfst to address the research problems suggested by my supervisor. He knew about my little project and help me improve the software by making new requirements and giving more suggestions on the design. He helped me with the manuscript drafting and revising, and encouraged me to submit it to JOSS. This work could never be completed without his contribution, as well as the great suggestions from the reviewers.
