### 2020-08-13 #2388 @mikldk
@tomsing1, @rcannood: Can you confirm that you have both finished the review?

@hope-data-science:

* Please make a tagged release and archive (e.g. with [Zenodo](https://zenodo.org/)) as described [here](<https://joss.readthedocs.io/en/latest/submitting.html#the-review-process>), and report the version number and archive DOI in this thread. Please verify that the archive deposit has the correct metadata (title and author list), or edit these if that is not the case.
