### 2020-08-14 #2388 @hope-data-science
Response:
1. Please have a final read though of the paper, checking language etc.
Reply: Done as suggested.

2. Have a final check of the proofs with @whedon generate pdf
Reply: Done as suggested.

3. Please make a tagged release and archive (e.g. with Zenodo) as described here, and report the version number and archive DOI in this thread. Please verify that the archive deposit has the correct metadata (title and author list), or edit these if that is not the case.
Reply: Done. For the archive, please find it in <https://zenodo.org/badge/latestdoi/240626994>.

If there are further questions, please let me know.

Thanks.
