### 2020-08-17 #2388 @hope-data-science
> @hope-data-science I'm not sure if it is v0.9.7 or v0.9.8 that you are aiming to publish? Please ensure that title and authors are the same on Zenodo and the paper.

@mikldk  When I submit the draft, the software is at the v0.9.7. But after the revision, I submit is again on CRAN and it leaps to v0.9.8. I am not sure which version should the draft belong, is there a routine way to provide it?
