### 2020-08-17 #2388 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
